+++ docs/project/changelog.rst
@@ -35,6 +35,10 @@ notice.
 Bug fixes
 .........
 
+* Fixed ``connection.recv(timeout=0)`` in the :mod:`threading` implementation.
+  If a message is already received, it is returned. Previously,
+  :exc:`TimeoutError` was raised incorrectly.
+
 * Wrapped errors when reading the opening handshake request or response in
   :exc:`~exceptions.InvalidMessage` so that :func:`~asyncio.client.connect`
   raises :exc:`~exceptions.InvalidHandshake` or a subclass when the opening